chore: rename internal package to pkg
