+++ .github/workflows/ci.yml
@@ -17,20 +17,20 @@ jobs:
         uses: actions/checkout@v4.2.2
         with:
           fetch-depth: 0
-      
+
       - name: Set up Go
         uses: actions/setup-go@v5
         with:
           go-version: '1.24'
           check-latest: true
           cache: true
-      
+
       - name: Lint
         uses: golangci/golangci-lint-action@v8
         with:
           version: latest
           args: --timeout=5m
-  
+
   test:
     name: Test
     runs-on: ubuntu-latest
@@ -39,27 +39,27 @@ jobs:
         uses: actions/checkout@v4.2.2
         with:
           fetch-depth: 0
-      
+
       - name: Set up Go
         uses: actions/setup-go@v5
         with:
           go-version: '1.24'
           check-latest: true
           cache: true
-      
+
       - name: Install dependencies
         run: go mod download
-      
+
       - name: Run tests with coverage
         run: go test -race -coverprofile=coverage.txt -covermode=atomic ./...
         continue-on-error: true
-      
+
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v5
         with:
           file: ./coverage.txt
           fail_ci_if_error: false
-          
+
       - name: Archive code coverage results
         uses: actions/upload-artifact@v4
         with:
@@ -76,21 +76,21 @@ jobs:
         uses: actions/checkout@v4.2.2
         with:
           fetch-depth: 0
-      
+
       - name: Set up Go
         uses: actions/setup-go@v5
         with:
           go-version: '1.24'
           check-latest: true
           cache: true
-      
+
       - name: Build binary
         run: |
-          go build -v -ldflags="-X github.com/Azure/mcp-kubernetes/internal/version.GitCommit=$(git rev-parse HEAD) -X github.com/Azure/mcp-kubernetes/internal/version.BuildMetadata=$(date +%Y%m%d)" -o mcp-kubernetes ./cmd/mcp-kubernetes
-      
+          go build -v -ldflags="-X github.com/Azure/mcp-kubernetes/pkg/version.GitCommit=$(git rev-parse HEAD) -X github.com/Azure/mcp-kubernetes/pkg/version.BuildMetadata=$(date +%Y%m%d)" -o mcp-kubernetes ./cmd/mcp-kubernetes
+
       - name: Build Docker image
         run: docker build -t mcp-kubernetes:test --build-arg VERSION=$(git describe --tags --always --dirty | cut -c2-),GIT_COMMIT=$(git rev-parse HEAD),BUILD_DATE=$(date +%Y%m%d),GIT_TREE_STATE=$(if git diff --quiet; then echo "clean"; else echo "dirty"; fi) .
-      
+
       - name: Check Docker image
         run: |
           docker images mcp-kubernetes:test
@@ -104,7 +104,7 @@ jobs:
         uses: actions/checkout@v4.2.2
         with:
           fetch-depth: 0
-      
+
       - name: Run Gosec Security Scanner
         uses: securego/gosec@master
         with:
