+++ Dockerfile
@@ -17,7 +17,7 @@ RUN go mod download
 COPY . .
 
 # Build the application
-RUN CGO_ENABLED=0 GOOS=linux go build -ldflags "-X github.com/Azure/mcp-kubernetes/internal/version.GitVersion=${VERSION} -X github.com/Azure/mcp-kubernetes/internal/version.GitCommit=${GIT_COMMIT} -X github.com/Azure/mcp-kubernetes/internal/version.BuildMetadata=${BUILD_DATE} -X github.com/Azure/mcp-kubernetes/internal/version.GitTreeState=${GIT_TREE_STATE}" -o mcp-kubernetes ./cmd/mcp-kubernetes
+RUN CGO_ENABLED=0 GOOS=linux go build -ldflags "-X github.com/Azure/mcp-kubernetes/pkg/version.GitVersion=${VERSION} -X github.com/Azure/mcp-kubernetes/pkg/version.GitCommit=${GIT_COMMIT} -X github.com/Azure/mcp-kubernetes/pkg/version.BuildMetadata=${BUILD_DATE} -X github.com/Azure/mcp-kubernetes/pkg/version.GitTreeState=${GIT_TREE_STATE}" -o mcp-kubernetes ./cmd/mcp-kubernetes
 
 # Runtime stage
 FROM alpine:3.19
