+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/Azure/mcp-kubernetes
 
-go 1.23
+go 1.24
 
 toolchain go1.24.2
