+++ pkg/cilium/registry.go
