+++ pkg/kubectl/executor.go
@@ -4,10 +4,10 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/Azure/mcp-kubernetes/internal/command"
-	"github.com/Azure/mcp-kubernetes/internal/config"
-	"github.com/Azure/mcp-kubernetes/internal/security"
-	"github.com/Azure/mcp-kubernetes/internal/tools"
+	"github.com/Azure/mcp-kubernetes/pkg/command"
+	"github.com/Azure/mcp-kubernetes/pkg/config"
+	"github.com/Azure/mcp-kubernetes/pkg/security"
+	"github.com/Azure/mcp-kubernetes/pkg/tools"
 )
 
 // KubectlExecutor implements the CommandExecutor interface for kubectl commands
