+++ pkg/kubectl/registry.go
